testing/web-platform/tests/DOMEvents/tests/submissions/Microsoft/converted/dispatchEvent.NOT_SUPPORTED_ERR.html

<!DOCTYPE html>
<html>
<head>
<title> dispatchEvent() DOMException.NOT_SUPPORTED_ERR </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>

<div style="display: none">
    <input id="target" type="checkbox">
    <button id="button1"> Cache the MouseEvent object </button>
    <button id="button2"> Dispatch the cached MouseEvent object </button>
</div>

<script>

    var EventObject;
    var TARGET = document.getElementById("target");
    var Button1 = document.getElementById("button1");
    var Button2 = document.getElementById("button2");

    var description = "Test Description: " +
                      "dispatchEvent - DOMException NOT_SUPPORTED_ERR raises if the Event object " +
                      "has not been created using DocumentEvent.createEvent().";

    var t = async_test(description);

    Button1.addEventListener("click", function(e){ EventObject = e }, true);
    Button2.addEventListener("click", t.step_func(DispatchEvent), true);

    window.onload = function()
    {
        Button1.click();
        Button2.click();
    }

    function DispatchEvent()
    {
        assert_throws("NOT_SUPPORTED_ERR", function()
        {
            TARGET.dispatchEvent(EventObject);
        });
        t.done();
    }
</script>
</body>
</html>
